api/requirements.txt
# EchoTuner API v1.0 Dependencies
# Core FastAPI and server
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.0

# HTTP and async support
httpx==0.28.1
requests==2.32.4
aiohttp==3.12.13

# AI and ML
ollama==0.5.1
numpy==2.3.1

# Spotify integration
spotipy==2.25.1

# Security (optional, for future auth features)
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
